metadata-ingestion/chrono/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
confluent-kafka==1.9.2
idna==3.3
parsec==3.13
protobuf==4.21.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
urllib3==1.26.12
